+++ Logging/targets/Email.ps1
@@ -7,6 +7,8 @@
         Subject     = @{Required = $false;  Type = [string]}
         Credential  = @{Required = $false;  Type = [pscredential]}
         Level       = @{Required = $false;  Type = [string]}
+        Port        = @{Required = $false;  Type = [int]}
+        UseSsl      = @{Required = $false;  Type = [bool]}
     }
     Logger = {
         param(
@@ -21,6 +23,8 @@
         $Params['From'] = $Configuration.From
         $Params['To'] = $Configuration.To.Split(',').Trim()
         $Params['Body'] = Replace-Token -String $Format -Source $Log
+        $Params['Port'] = $Configuration.Port
+        $Params['UseSsl'] = $Configuration.UseSsl
 
         if ($Configuration.Subject) {
             $Params['Subject'] = Replace-Token -String $Configuration.Subject -Source $Log
@@ -30,6 +34,14 @@
 
         if ($Configuration.Credential) {
             $Params['Credential'] = $Configuration.Credential
+        }
+
+        if ($Configuration.Port) {
+            $Params['Port'] = $Configuration.Port
+        }
+
+        if ($Configuration.UseSsl) {
+            $Params['UseSsl'] = $Configuration.UseSsl
         }
 
         if ($Log.Body) {