+++ Tests/Logging.Tests.ps1
@@ -153,13 +153,13 @@ InModuleScope Logging {
     Describe 'Logging Targets' {
         It 'loads the logging targets' {
             $Targets = $InitialSessionState.Variables.Item('LogTargets').Value
-            $Targets.Count | Should Be 5
+            $Targets.Count | Should Be 6
         }
 
         It 'returns the loaded logging targets' {
             $AvailableTargets = Get-LoggingTargetAvailable
             $AvailableTargets | Should Be System.Collections.Hashtable+SyncHashtable
-            $AvailableTargets.Count | Should Be 5
+            $AvailableTargets.Count | Should Be 6
         }
     }