+++ build.ps1
@@ -12,4 +12,5 @@ Import-Module Psake, BuildHelpers, platyPS, PSScriptAnalyzer
 Set-BuildEnvironment
 
 Invoke-psake -buildFile .\build.psake.ps1 -taskList $Task -nologo
-exit ([int](-not $psake.build_success))
+
+exit ([int] (-not $psake.build_success))